Deprecate DRM types Make CancellableResult equatable
