+++ r2-shared-swift/Toolkit/CancellableResult.swift
@@ -105,3 +105,6 @@ public enum CancellableResult<Success, Failure: Error> {
     }
 
 }
+
+extension CancellableResult: Equatable where Success: Equatable, Failure: Equatable {
+}
